Modularize the policy controller by moving inbound and outbound code into seperate modules
