+++ policy-controller/k8s/index/src/inbound/mod.rs
@@ -1,0 +1,9 @@
+pub mod authorization_policy;
+mod http_route;
+pub mod index;
+mod meshtls_authentication;
+mod network_authentication;
+mod server;
+pub mod server_authorization;
+
+pub use index::{Index, SharedIndex};
