+++ policy-controller/k8s/index/src/outbound/mod.rs
@@ -1,0 +1,3 @@
+pub mod index;
+
+pub use index::{Index, ServiceRef, SharedIndex};
