+++ policy-controller/src/admission.rs
@@ -427,7 +427,7 @@ impl Validate<ServerAuthorizationSpec> for Admission {
 #[async_trait::async_trait]
 impl Validate<HttpRouteSpec> for Admission {
     async fn validate(self, _ns: &str, _name: &str, spec: HttpRouteSpec) -> Result<()> {
-        use index::http_route::convert;
+        use index::http_route;
 
         fn validate_match(
             httproute::HttpRouteMatch {
@@ -437,18 +437,18 @@ impl Validate<HttpRouteSpec> for Admission {
                 method,
             }: httproute::HttpRouteMatch,
         ) -> Result<()> {
-            let _ = path.map(convert::path_match).transpose()?;
+            let _ = path.map(http_route::path_match).transpose()?;
             let _ = method
                 .as_deref()
                 .map(core::http_route::Method::try_from)
                 .transpose()?;
 
             for q in query_params.into_iter().flatten() {
-                convert::query_param_match(q)?;
+                http_route::query_param_match(q)?;
             }
 
             for h in headers.into_iter().flatten() {
-                convert::header_match(h)?;
+                http_route::header_match(h)?;
             }
 
             Ok(())
@@ -458,9 +458,9 @@ impl Validate<HttpRouteSpec> for Admission {
             match filter {
                 httproute::HttpRouteFilter::RequestHeaderModifier {
                     request_header_modifier,
-                } => convert::req_header_modifier(request_header_modifier).map(|_| ()),
+                } => http_route::req_header_modifier(request_header_modifier).map(|_| ()),
                 httproute::HttpRouteFilter::RequestRedirect { request_redirect } => {
-                    convert::req_redirect(request_redirect).map(|_| ())
+                    http_route::req_redirect(request_redirect).map(|_| ())
                 }
             }
         }
